python3: disable tkinter module on posix (#34998)
When running with VCPKG_FORCE_SYSTEM_BINARIES=ON (linux/arm64) then the Python
build will pick up libtk and attempt to build the tkinter module, linking to the
system library and causing either a build failure or an external link
dependency. Fix this by explicitly disabling it as we do for curses.