[libmagic] Explicitly toggle features. (#37459)
Libmagic supports some compression libraries that by default are enabled
unpredictably, if they are found in the system. This PR adds features to
the port that enable support for these compression libraries, and makes
sure the features are disabled if not specified.

Validated by building the port with all features enabled on Windows and
Linux, and importing it in a project with static library linkage on
Windows.