[vcpkg-qmake] Remove unnecessary directory (#37752)
Recently when I was using vcpkg-qmake, I found that a function in the
`ports/vcpkg-qmake/vcpkg_qmake_configure.cmake` file was repeated:

https://github.com/microsoft/vcpkg/blob/be9eb6694585a8db137872b29fcad3619a5f893a/ports/vcpkg-qmake/vcpkg_qmake_configure.cmake#L143
By running this line of code, we will configure the generated `qt.conf`
file. At the same time, if the
`${CURRENT_BUILDTREES_DIR}/${config_triplet}/` path does not exist, we
will also generate it.
Then we back up the environment variables and run the code
https://github.com/microsoft/vcpkg/blob/be9eb6694585a8db137872b29fcad3619a5f893a/ports/vcpkg-qmake/vcpkg_qmake_configure.cmake#L149
to generate the existing path above.